auditwheel choose higher priority tag when possible
